migrate is_floating_point to tensor.py
